Upload the video and video details complete with user authentication and authorization

+++ src/controller/video.controller.js
@@ -2,8 +2,8 @@ const fs = require("node:fs/promises");
 const path = require("node:path");
 const crypto = require("node:crypto");
 const { deleteFolder, deleteFile } = require("../utils");
-const { pipeline } = require('node:stream/promises');
-
+const { pipeline } = require("node:stream/promises");
+const Video = require("../model/video.schema");
 
 const uploadVideoController = async (req, res) => {
   const fileName = req.headers.filename || "untitled";
@@ -30,6 +30,25 @@ const uploadVideoController = async (req, res) => {
     // req.pipe(fileStream);
     await pipeline(req, fileStream);
 
+    // update the database after the video is uploaded successfully
+    // await Video.create({});
+
+    const video = new Video({
+      videoId,
+      name,
+      user: req?.user.id,
+      extension,
+      dimensions: {
+        width: 1080,
+        height: 720,
+      },
+      extractedAudio: false,
+      resizes: [],
+    });
+
+    // await video.save();
+
+    file.close();
     return res.status(200).send("File uploaded successfully");
   } catch (error) {
     // if the error is because the folder does not exist
@@ -41,7 +60,12 @@ const uploadVideoController = async (req, res) => {
 
     // if the error is on pipelining i.e user cancels the upload process
     // in that case delete the folder that was created and remove any unnecessary files.
-    await deleteFolder(`./storage/${videoId}`);
+
+    if (error.code === "ECONNRESET") {
+      await deleteFolder(`./storage/${videoId}`);
+    }
+
+    console.log("Error on uploading file", error);
   }
 };
